add UpdateGame method - NOT WORKING
I don't understand why this isn't working. I think I followed tutorial as far as this file goes.

Game object is updated within EditGame file when ADDING a game, but not when UPDATING a game.

Game name does not update on submit, therefore new name is never passed out of this file.